requirements.txt
python-telegram-bot==21.1.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
nest_asyncio==1.6.0
pandas==2.2.2
openpyxl==3.1.2
apscheduler==3.10.4